+++ Cargo.toml
@@ -24,3 +24,4 @@ sentry = "0.12.0"
 log = "0.4"
 pretty_env_logger = "0.3"
 byte-unit = "2.1.0"
+futures = "0.1"
