+++ src/lib.rs
@@ -19,3 +19,9 @@ pub use parquet_writer::*;
 
 mod cleaner;
 pub use cleaner::*;
+
+pub mod covid19_data;
+pub use covid19_data::*;
+
+pub mod world_population;
+pub use world_population::*;
